Fixed query optimization.
